+++ sites/all/libraries/fontello/lib/system/init/bundle/styles.js
@@ -7,9 +7,9 @@
 
 
 // stdlib
-var fs    = require('fs');
-var path  = require('path');
-var crypto = require('crypto');
+var fs      = require('fs');
+var path    = require('path');
+var assert  = require('assert');
 
 
 // 3rd-party
@@ -21,86 +21,34 @@ var fstools = require('fs-tools');
 var stopwatch = require('../utils/stopwatch');
 var RENDERERS = require('./styles/renderers');
 var findPaths = require('./utils/find_paths');
+var Processor = require('./utils/cached_processor');
 
 
 ////////////////////////////////////////////////////////////////////////////////
-
-// dummy helper to generate md5 hash of a `string`
-//
-function strDigest(str) {
-  return crypto.createHash('md5').update(str).digest('hex');
-}
 
 // Compile styles for all blocks in package (from all files)
 //
 function joinStyles(lookup, destination, options) {
   var result = [];
 
-  // calculated dile digest (md5)
-  var fileDigest = _.memoize(function (file) {
-    var content = '';
-    try { content = fs.readFileSync(file, 'utf-8'); } catch(e) {}
-    return strDigest(content);
+  var processor = new Processor({
+    cache: path.join(options.cacheDir, options.pkgName + '.json')
   });
 
-  findPaths(lookup, function (file) {
+  processor.process = function (file) {
     var extname = path.extname(file)
       , renderer = RENDERERS[extname];
-    
-    if (!renderer) {
-      throw new Error('Don\'t know how to compile ' + file);
-    }
 
-    var cacheFilename = path.join(options.cacheDir, strDigest(file)) + '.json';
+    assert.ok(renderer, 'Don\'t know how to compile ' + file);
 
-    // Cache structure:
-    //
-    // {
-    //   resultCSS: 'compiled css',
-    //   dependencies: {
-    //     'dependency.styl': 'dependency digest',
-    //     ...
-    //   }
-    // }
-    //
-    // NOTE: Dependencies include rendered file and all imported files
-    //
-    var cache = {
-      resultCSS: null
-    , dependencies: {}
-    };
+    var data = renderer(file, options);
+    processor.addDependencies(file, data.imports);
 
-    // Load cache, if exists
-    try { cache = JSON.parse(fs.readFileSync(cacheFilename, 'utf-8')); } catch(e) {}
+    return data.css;
+  };
 
-    // Check if file or dependencies changed (calculate signatures)
-    if ((null !== cache.resultCSS) &&
-        _.any(cache.dependencies, function(digest, file) {
-          return digest !== fileDigest(file);
-        })
-      )
-    {
-      cache.resultCSS = null;
-    }
-
-    // rebuild if no valid cache
-    if (null === cache.resultCSS) {
-      // Invoke renderer
-      var data = renderer(file, options);
-
-      cache.resultCSS = data.css;
-
-      // Update dependencies
-      cache.dependencies = _.transform(data.imports, function(dependencies, file){
-        dependencies[file] = fileDigest(file);
-      }, {});
-
-      // Save cache
-      fstools.mkdirSync(path.dirname(cacheFilename));
-      fs.writeFileSync(cacheFilename, JSON.stringify(cache));
-    }
-
-    result.push(cache.resultCSS);
+  findPaths(lookup, function (file) {
+    result.push(processor.get(file));
   });
 
   fstools.mkdirSync(path.dirname(destination));
@@ -111,11 +59,10 @@ function joinStyles(lookup, destination, options) {
 ////////////////////////////////////////////////////////////////////////////////
 
 
-module.exports = function (sandbox, callback) {
+module.exports = function (sandbox) {
   var N      = sandbox.N
     , tmpdir = sandbox.tmpdir
     , cacheDir = path.join(N.config.options.cache_dir, 'modules_styles')
-    , err = null
     , timer  = stopwatch();
 
   _.keys(sandbox.config.packages).forEach(function (pkgName) {
@@ -132,52 +79,45 @@ module.exports = function (sandbox, callback) {
       return;
     }
 
-    try {
-      joinStyles(stylesConfig, stylesTmpFile, {
-        pkgName:  pkgName
-      , packages: sandbox.config.packages
-      , cacheDir: cacheDir
-      });
+    joinStyles(stylesConfig, stylesTmpFile, {
+      pkgName:  pkgName
+    , packages: sandbox.config.packages
+    , cacheDir: cacheDir
+    });
 
-      // Use `main` file, if exists. Otherwise use concatenated styles file only.
-      if (_.find(stylesConfig, 'main')) {
-        targetFile = _.find(stylesConfig, 'main').main;
-      } else {
-        targetFile = stylesTmpFile;
-      }
-
-      // Prepend path with styles tree to allow use
-      //
-      //    //= require styles
-      //
-      // in main file.
-      environment.prependPath(stylesTmpDir);
-
-      // When Mincer is asked for a main file, it must be within roots, that
-      // Mincer knows about. See: https://github.com/nodeca/mincer/issues/51
-      stylesConfig.forEach(function (options) {
-        environment.appendPath(options.root);
-      });
-
-      // Find & build asset
-      var asset = environment.findAsset(targetFile);
-
-      // Check that main file is requirable.
-      if (!asset) {
-        // Restore Mincer's paths.
-        environment.clearPaths();
-        environment.appendPath(originPaths);
-
-        err = 'Main style file of ' + pkgName + ' not found: ' + targetFile;
-        return false;
-      }
-
-      // Write main file.
-      fs.writeFileSync(resultFile, asset.buffer, 'utf8');
-    } catch (e) {
-      err = e;
-      return false;
+    // Use `main` file, if exists. Otherwise use concatenated styles file only.
+    if (_.find(stylesConfig, 'main')) {
+      targetFile = _.find(stylesConfig, 'main').main;
+    } else {
+      targetFile = stylesTmpFile;
     }
+
+    // Prepend path with styles tree to allow use
+    //
+    //    //= require styles
+    //
+    // in main file.
+    environment.prependPath(stylesTmpDir);
+
+    // When Mincer is asked for a main file, it must be within roots, that
+    // Mincer knows about. See: https://github.com/nodeca/mincer/issues/51
+    stylesConfig.forEach(function (options) {
+      environment.appendPath(options.root);
+    });
+
+    // Find & build asset
+    var asset = environment.findAsset(targetFile);
+
+    // Check that main file is requirable.
+    if (!asset) {
+      // Restore Mincer's paths.
+      environment.clearPaths();
+      environment.appendPath(originPaths);
+      throw new Error('Main style file of ' + pkgName + ' not found: ' + targetFile);
+    }
+
+    // Write main file.
+    fs.writeFileSync(resultFile, asset.buffer, 'utf8');
 
     // restore mincer's paths
     environment.clearPaths();
@@ -188,5 +128,4 @@ module.exports = function (sandbox, callback) {
   });
 
   N.logger.info('Processed styles section %s', timer.elapsed);
-  callback(err);
 };